+++ .github/workflows/test.yaml
@@ -59,68 +59,11 @@ jobs:
                   PLATFORM: ${{ matrix.os }}
                   DISPLAY: :42
               run: |
-                  uv run pytest --cov --color=yes --cov-report=xml --ignore=tests/io/remote_storage/
+                  uv run pytest --cov --color=yes --cov-report=xml
             - name: Upload coverage to Codecov
               uses: codecov/codecov-action@v5
               with:
                   name: coverage
-                  verbose: true
-              env:
-                  CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
-
-    test-remote-storage:
-        runs-on: ubuntu-latest
-        defaults:
-            run:
-                shell: bash
-        strategy:
-            fail-fast: false
-            matrix:
-                python: ["3.11", "3.13"]
-        env:
-            MPLBACKEND: agg
-            PLATFORM: ubuntu-latest
-            DISPLAY: :42
-            GCSFS_EXPERIMENTAL_ZB_HNS_SUPPORT: "false"
-        steps:
-            - uses: actions/checkout@v6
-            - uses: astral-sh/setup-uv@v7
-              with:
-                  version: "latest"
-                  python-version: ${{ matrix.python }}
-            - name: Install dependencies
-              run: |
-                  uv add dask
-                  uv sync --group=test
-            - name: Build and start storage emulators
-              run: |
-                  docker build -f tests/io/remote_storage/Dockerfile.emulators -t spatialdata-emulators .
-                  docker run --rm -d --name spatialdata-emulators \
-                    -p 5000:5000 -p 10000:10000 -p 4443:4443 \
-                    spatialdata-emulators
-            - name: Wait for emulator ports
-              run: |
-                  echo "Waiting for S3 (5000), Azure (10000), GCS (4443)..."
-                  python3 -c "
-                  import socket, time
-                  for _ in range(45):
-                      try:
-                          for p in (5000, 10000, 4443):
-                              socket.create_connection(('127.0.0.1', p), timeout=2)
-                          print('Emulators ready.')
-                          break
-                      except (socket.error, OSError):
-                          time.sleep(2)
-                  else:
-                      raise SystemExit('Emulators did not become ready.')
-                  "
-            - name: Test remote storage
-              run: |
-                  uv run pytest tests/io/remote_storage/ --cov --color=yes --cov-report=xml
-            - name: Upload coverage to Codecov
-              uses: codecov/codecov-action@v5
-              with:
-                  name: coverage-remote-storage-${{ matrix.python }}
                   verbose: true
               env:
                   CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
