--- tests/io/remote_storage/Dockerfile.emulators
@@ -1,33 +1,0 @@
-# Storage emulators for tests in this directory (S3, Azure, GCS).
-# Emulator URLs: S3 127.0.0.1:5000 | Azure 127.0.0.1:10000 | GCS 127.0.0.1:4443
-#
-# Build (from project root):
-#   docker build -f tests/io/remote_storage/Dockerfile.emulators -t spatialdata-emulators .
-#
-# Run in background (detached):
-#   docker run --rm -d --name spatialdata-emulators -p 5000:5000 -p 10000:10000 -p 4443:4443 spatialdata-emulators
-#
-# Run in foreground (attach to terminal):
-#   docker run --rm --name spatialdata-emulators -p 5000:5000 -p 10000:10000 -p 4443:4443 spatialdata-emulators
-#
-# Stop / remove:
-#   docker stop spatialdata-emulators
-#   docker rm -f spatialdata-emulators   # if already stopped or to force-remove
-FROM node:20-slim
-RUN apt-get update && apt-get install -y --no-install-recommends \
-    python3 python3-pip python3-venv curl ca-certificates \
-    && rm -rf /var/lib/apt/lists/*
-RUN npm install -g azurite
-RUN python3 -m venv /opt/venv && /opt/venv/bin/pip install --no-cache-dir 'moto[server]'
-ENV PATH="/opt/venv/bin:$PATH"
-# fake-gcs-server must match the image CPU. `ARG TARGETARCH=amd64` can stay amd64 on arm64 builds.
-RUN set -eux; \
-    arch="$(uname -m)"; \
-    case "$arch" in x86_64) fgs=amd64 ;; aarch64|arm64) fgs=arm64 ;; *) echo "unsupported arch: $arch" >&2; exit 1 ;; esac; \
-    cd /tmp && curl -fsSL -o fgs.tgz "https://github.com/fsouza/fake-gcs-server/releases/download/v1.54.0/fake-gcs-server_1.54.0_linux_${fgs}.tar.gz" \
-    && tar xzf fgs.tgz && mv fake-gcs-server /usr/local/bin/ 2>/dev/null || mv fake-gcs-server_*/fake-gcs-server /usr/local/bin/ \
-    && chmod +x /usr/local/bin/fake-gcs-server && rm -f fgs.tgz
-RUN mkdir -p /data
-EXPOSE 5000 10000 4443
-RUN echo 'moto_server -H 0.0.0.0 -p 5000 & azurite --silent --location /data --blobHost 0.0.0.0 --skipApiVersionCheck & fake-gcs-server -scheme http -port 4443 & wait' > /start.sh && chmod +x /start.sh
-CMD ["/bin/sh", "/start.sh"]
